+++ src/components/common/Calendar/Calendar.module.scss
@@ -1,3 +1,5 @@
+
+@import '@/styles/theme.scss';
 // 최상단 컨테이너
 :global(.react-calendar) {
   border: none;
@@ -6,12 +8,12 @@
 
 // 날짜 텍스트 컬러 
 :global(.react-calendar__month-view__days__day) {
-  color :#7D8691
+  color :$gray500;
 }
 
 // 오늘 날짜를 hover
 :global(.react-calendar__tile--now:enabled:hover) {
-  background-color: #e6e6e6;
+  background-color: #e6e6e6; //호버 컬러는 임의로 지었어요. 어플이라서 TODO: 디자인 컬러 받기
 }
 
 // 날짜를 hover했을때
@@ -22,7 +24,7 @@
 
 // 날짜를 클릭해서 focused 상태인경우
 :global(.react-calendar__tile--active:enabled:focus) {
-  background-color: #8B95A1;
+  background-color: $gray500; // 400
 }
 
 
@@ -34,21 +36,21 @@
 
 // today active 날짜
 :global(.react-calendar__tile--active) {
-  background-color: #8B95A1;
+  background-color: $gray500; // 400
   border-radius: 8px;
   color: #fff;
 }
 
 // 날짜가 오늘보다 이전인 날짜 
 :global(.past-date) {
-  color: #DBDEE2;
+  color: $gray300;
   text-decoration: line-through;
   pointer-events: none;
 }
 
 // [일,월,화,수] 표시 요일
 :global(.react-calendar__month-view__weekdays) {
-  color: #7D8691;
+  color: $gray500;
   font-size: 13px;
   text-decoration: none;
 }
@@ -58,4 +60,12 @@
   > img,svg {
     transform: rotate(180deg); /* 화살표를 180도 돌립니다. */
   }
+}
+
+.PrevArrowIcon {
+  stroke: $gray500; //400
+}
+
+.NextArrowIcon {
+  stroke: $black;
 }
