+++ src/styles/theme.scss
@@ -6,3 +6,4 @@ $gray700: #687079;
 $gray800: #535961;
 $gray1000: #313438;
 $primary: #383cd1;
+$black: #000000;
